+++ CHANGELOG.md
@@ -14,6 +14,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed the 'filters' count column in [Catalogue] tree collection showing edit control when clicked
 - Fixed Find not working when searching by ID for [Pipeline] objects
 
+### Added
+
+- Added ability to filter [Catalogue] in the Find dialog by Internal/Deprecated etc
+
+
 ## [7.0.3] - 2021-11-04
 
 ### Fixed